requirements.txt
playwright==1.41.2
pandas==2.2.0
schedule==1.2.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
google-cloud-storage==2.12.0
